# openapi.yaml
openapi: 3.0.3
info:
  title: Releasarr API
  version: 0.1.0
  description: |
    Contract derived from the new frontend client. Captures the HTTP interface the
    UI expects today so the backend can be validated and mocked consistently.
  license:
    name: Proprietary
    url: https://github.com/koreetz/releasarr
servers:
  - url: https://api.releasarr.dev/api
    description: Production server
  - url: http://localhost:8001/api
    description: Local development server
security:
  - BearerAuth: []
  - ServiceKeyAuth: []
paths:
  /requests:
    get:
      tags: [Requests]
      summary: List media requests
      operationId: listRequests
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/RequestStatus'
        - $ref: '#/components/parameters/RequestType'
        - $ref: '#/components/parameters/RequestOwner'
      responses:
        '200':
          description: Paged list of media requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestsResponse'
        '400':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is restricted to their own requests and may not filter by owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Requests]
      summary: Create a media request
      operationId: createRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaRequestCreate'
      responses:
        '201':
          description: Created request details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '400':
          description: Malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /requests/{requestId}:
    get:
      tags: [Requests]
      summary: Retrieve a single request
      operationId: getRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Requested media request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Requests]
      summary: Partially update a media request
      operationId: updateRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaRequestUpdate'
      responses:
        '200':
          description: Updated request details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '400':
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for at least one field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Requests]
      summary: Delete a media request
      operationId: deleteRequest
      description: |
        Removes the request and unmonitors what produced it: the season in Sonarr
        or the movie in Radarr. Dropping the row alone would not last, because the
        recurring sync rebuilds a request for every monitored season Sonarr still
        reports as missing.

        The series or movie itself stays in the library, as do any files already
        imported, and a series stays monitored even once its last season has
        gone - Sonarr reads one with no monitored season as wanting nothing.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '204':
          description: Request removed.
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr or Radarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /requests/{requestId}/episodes:
    get:
      tags: [Requests]
      summary: List the episodes of the season a request covers
      operationId: listRequestEpisodes
      description: |
        Every episode of the one season a series request covers, with its air
        date and the file Sonarr holds for it, if any. Sonarr is asked directly
        rather than the answer being derived from the files of the releases
        grabbed here: a season is just as likely to have been filled from
        outside releasarr, and what Sonarr holds is the truth either way.

        An episode counts as `downloaded` once Sonarr has a file for it,
        whatever its air date says. Of the rest, one whose air date has passed
        is `missing` - the only state that is anybody's to act on - and one
        still to air, or that Sonarr has no date for at all, is `unaired`.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The episodes of the request's season, in episode order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonEpisodesResponse'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request has no season in Sonarr whose episodes can be listed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /requests/{requestId}/seasons:
    get:
      tags: [Requests]
      summary: List the seasons of the series a request belongs to
      operationId: listRequestSeasons
      description: |
        The seasons of the series behind a series request, with the request state
        of each, for managing the selection from the request itself. Sonarr is
        asked by the series id the sync stamped on the request, which is the only
        handle a request page has - it knows nothing about TVDB. The monitored
        flags come from Sonarr itself, so a season monitored there without ever
        going missing is reported as monitored here too.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The seasons of the request's series.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesSeasonsResponse'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request has no series in Sonarr whose seasons can be managed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Requests]
      summary: Set which seasons of a series Sonarr monitors
      operationId: updateRequestSeasons
      description: |
        Brings the monitoring of a series in line with the given selection,
        writing the flags straight to Sonarr. The difference is taken against
        what Sonarr monitors now rather than against the requests releasarr
        holds, because a monitored season that is already complete never becomes
        a request and would otherwise be unmonitored on save.

        Newly picked seasons are monitored and synced into requests; seasons
        dropped from the selection are unmonitored and any request of theirs
        removed, which includes the request in the path if its own season is left
        out.

        Both halves go through a single Sonarr write, so the series never passes
        through a state where every season has been dropped and is about to be
        added back.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeasonsPayload'
      responses:
        '200':
          description: The seasons of the request's series, as they now stand.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesSeasonsResponse'
        '400':
          description: The series has no such season.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request has no series in Sonarr whose seasons can be managed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /discover/requests:
    post:
      tags: [Discover]
      summary: Add media to Sonarr/Radarr and request it
      description: |
        Adds the picked series or movie to Sonarr/Radarr by its TVDB/TMDB id,
        monitoring the selected seasons, and creates the matching media requests
        before responding. Media already in the library is not re-added: only the
        monitoring is widened to cover the selection.
      operationId: addRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRequestPayload'
      responses:
        '201':
          description: The created requests, one per season for a series.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddRequestResponse'
        '400':
          description: Invalid root folder, season selection or quality profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No media matches the provider id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr or Radarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token obtained from /auth/login or /auth/refresh.
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        A service API key, acting as the user it is bound to. May be combined
        with X-Act-As-User to act as a different user, when the key allows it.
  parameters:
    RequestId:
      in: path
      name: requestId
      required: true
      schema:
        type: string
      description: Unique identifier for a media request.
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
      description: Results page to retrieve (1-indexed).
    PerPage:
      in: query
      name: per_page
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Number of items to return per page.
    RequestStatus:
      in: query
      name: status
      schema:
        $ref: '#/components/schemas/MediaRequestStatus'
      description: Filter requests by lifecycle status.
    RequestType:
      in: query
      name: type
      schema:
        $ref: '#/components/schemas/MediaType'
      description: Filter requests by media type.
    RequestOwner:
      in: query
      name: owner
      schema:
        type: string
      description: >-
        Filter requests by owner user id. Only callers with view_all_requests
        (or an admin) may use this; others are restricted to their own.
...
  schemas:
    MediaType:
      type: string
      enum: [movie, series]
    MediaRequestStatus:
      type: string
      enum: [pending, searching, downloading, completed, failed]
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error identifier.
        message:
          type: string
          description: Human-readable explanation of the error.
        details:
          description: Optional key/value data with context for the failure.
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
              - type: array
                items:
                  type: string
    RequestsResponse:
      type: object
      required: [requests, total, page, per_page]
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/MediaRequest'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
...
    SeasonEpisodesResponse:
      type: object
      required: [season_number, episodes]
      properties:
        season_number:
          type: integer
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/SeasonEpisode'
...
    MediaRequest:
      oneOf:
        - $ref: '#/components/schemas/MovieRequest'
        - $ref: '#/components/schemas/SeriesRequest'
      discriminator:
        propertyName: type
        mapping:
          movie: '#/components/schemas/MovieRequest'
          series: '#/components/schemas/SeriesRequest'
...
    MediaRequestCreate:
      oneOf:
        - $ref: '#/components/schemas/CreateMovieRequest'
        - $ref: '#/components/schemas/CreateSeriesRequest'
      discriminator:
        propertyName: type
        mapping:
          movie: '#/components/schemas/CreateMovieRequest'
          series: '#/components/schemas/CreateSeriesRequest'
    MediaRequestUpdate:
      type: object
      properties:
        title:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        poster_url:
          type: string
          nullable: true
        overview:
          type: string
          nullable: true
        genres:
          type: array
          items:
            type: string
          nullable: true
        status:
          $ref: '#/components/schemas/MediaRequestStatus'
        runtime:
          type: integer
          nullable: true
        imdb_id:
          type: string
          nullable: true
        season_number:
          type: integer
          nullable: true
        total_episodes:
          type: integer
          nullable: true
        series_title:
          type: string
          nullable: true
        series_year:
          type: integer
          nullable: true
        localizations:
          type: object
          nullable: true
          description: Replace or clear localized metadata when provided.
          additionalProperties:
            $ref: '#/components/schemas/MediaLocalization'
        owner_user_id:
          type: string
          nullable: true
          description: Reassign the request's owner. Admin only.
...
    SeriesSeasonsResponse:
      type: object
      required: [seasons]
      properties:
        tvdb_id:
          type: integer
          nullable: true
          description: >-
            Absent when the seasons were read from a series in the library rather
            than looked up by TVDB id, which is how managing a request arrives.
        in_library:
          type: boolean
          default: false
        library_id:
          type: integer
          nullable: true
        monitor_new_seasons:
          type: boolean
          default: false
          description: >-
            Whether Sonarr monitors seasons announced after the series was added.
            Only meaningful in the library; Sonarr has nowhere to record it until
            then.
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/SeasonOption'
...
    AddRequestResponse:
      type: object
      required: [requests]
      properties:
        requests:
          type: array
          description: One request per requested season for a series, a single one for a movie.
          items:
            $ref: '#/components/schemas/MediaRequest'
...